added toggle for grades list
